norme, attempt #1

+++ src/parser/utils.c
@@ -18,7 +18,7 @@ int			mandatory_reserv_word(int expect_type)
 
 	if (!g_parser.token)
 		return (0);
-	if (g_parser.token->type == expect_type)
+	if ((int)g_parser.token->type == expect_type)
 		return (expect_type);
 	if ((type = is_reserved_word(g_parser.token->value->str))
 	&& type == expect_type)
